previous/requirements.txt
absl-py==0.12.0
addict==2.4.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
catkin-pkg==0.4.23
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
cython==0.29.22
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
docutils==0.17
entrypoints==0.3
freetype-py==2.2.0
future==0.18.2
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.10
imageio==2.9.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
markdown==3.3.4
markupsafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.2
notebook==6.3.0
numpy==1.19.2
oauthlib==3.1.0
open3d==0.12.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
plyfile==0.7.3
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==2.7.9
pycollada==0.6
pycparser==2.20
pyglet==1.5.15
pygments==2.8.1
pyopengl==3.1.0
pyparsing==2.4.7
pyrender==0.1.45
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-ignite==0.4.4
pytz==2021.1
pywavelets==1.1.1
pyyaml==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
send2trash==1.5.0
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.31
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
trimesh==3.9.10
urdfpy==0.0.22
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
widgetsnbextension==3.5.1
pykdtree
